web-app/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Database
sqlalchemy>=2.0
aiosqlite>=0.19.0

# Gmail integration
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# OpenAI for classification
openai>=1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0
python-dateutil>=2.8.2

# Export/Reporting
openpyxl>=3.1.0
pandas>=2.0.0

# Development
pytest>=7.0
